Dodanie możliwości uruchamiania testów algorytmu w cpp na cuda

+++ em_alg_cpp_cuda/src/cpu/ProgramArgumentsParser.h
@@ -19,7 +19,7 @@ public:
     unsigned getBlocks();
     unsigned getPopulationSize();
     unsigned getIterationsCount();
-    unsigned getNeighborhoodDistance();
+    double getNeighborhoodDistance();
 
 private:
     void addOptions();
